backend/requirements.txt
fastapi
uvicorn
pymysql
pydantic
sqlalchemy

mangum
boto3

python-dotenv

python-jose
email-validator
passlib==1.7.4
bcrypt==4.0.1

python-multipart

openpyxl